+++ example/lib/widgets/download_node_row.dart
@@ -1,7 +1,6 @@
+import 'package:example/mocks/fake_http_download.dart';
 import 'package:example/widgets/syc_helper.dart';
 import 'package:flutter/material.dart';
-
-import '../mocks/fake_http_download.dart';
 
 class DownloadNodeRow extends StatelessWidget {
   final FakeHttpDownloadLeaf node;