+++ example/lib/widgets/sync_child_row.dart
@@ -1,8 +1,7 @@
+import 'package:example/mocks/fake_firebase.dart';
 import 'package:example/widgets/syc_helper.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_sync_tree/flutter_sync_tree.dart';
-
-import '../mocks/fake_firebase.dart';
 
 class SyncChildRow extends StatelessWidget {
   final SyncNode node;